stripe markdown private wikis

+++ app/controllers/wikis_controller.rb
@@ -1,23 +1,30 @@
 class WikisController < ApplicationController
-	before_filter :authenticate_user!
+  include ApplicationHelper
+
+  before_filter :authenticate_user!
+
+  after_action :verify_authorized, :except => :index
 
   def index
-    
+    @wikis = policy_scope(Wiki)
   end
 
   def show
     @wiki = Wiki.find(params[:id])
+    authorize @wiki
   end
 
   def new
     @user = current_user
     @wiki = Wiki.new
+    authorize @wiki
   end
 
   def create
     @user = current_user
-    @wiki = Wiki.new
+    @wiki = Wiki.new(wiki_params)
     @wiki.user = current_user
+    authorize @wiki
 
     if @wiki.save
       redirect_to @wiki, notice: "Wiki was saved successfully."
@@ -30,12 +37,15 @@ class WikisController < ApplicationController
   def edit
     @wiki = Wiki.find(params[:id])
     @users = User.all
+    @collaborator = Collaborator.new
+    authorize @wiki
   end
 
   def update
     @wiki = Wiki.find(params[:id])
+    authorize @wiki
 
-    # @wiki.assign_attributes(wiki_params)
+    @wiki.assign_attributes(wiki_params)
     @wiki.user_ids = params[:wiki][:user_ids] if params[:wiki][:user_ids].present?
 
     if @wiki.save
@@ -49,7 +59,7 @@ class WikisController < ApplicationController
 
   def destroy
     @wiki = Wiki.find(params[:id])
-    # authorize @wiki
+    authorize @wiki
 
     if @wiki.destroy
       flash[:notice] = "\"#{@wiki.title}\" was deleted successfully."
@@ -59,4 +69,11 @@ class WikisController < ApplicationController
       render :show
     end
   end
+
+ private
+
+  def wiki_params
+    params.require(:wiki).permit(:title, :body, :private, :user, :user_ids)
+  end
+
 end
